requirements.txt
google-genai
tree-sitter==0.20.2
tree_sitter_languages==1.10.2